requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crontab==1.0.1
cryptography==41.0.2
flower==2.0.0
frozenlist==1.4.0
greenlet==2.0.2
humanize==4.7.0
idna==3.4
kombu==5.3.1
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
paramiko==3.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycparser==2.21
PyNaCl==1.5.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
sentry-sdk==1.28.1
six==1.16.0
SQLAlchemy==2.0.19
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
YooMoney==0.1.0
